add fpat prompt for `grida code` command
