+++ cli/init/index.ts
@@ -284,3 +284,13 @@ function create_grida_fallback_dir(
     path: dir,
   };
 }
+
+export * from "./init-.env";
+export * from "./init-.gitignore";
+export * from "./init-.grida";
+export * from "./init-base-project-with-teplate";
+export * from "./init-config-designsource";
+export * from "./init-config-designsource-figma";
+export * from "./init-grida.config.js";
+export * from "./init-package.json";
+export * from "./init-pubspec.yaml";
